backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow
numpy
Pillow
google-generativeai
python-dotenv
opencv-python-headless